chore(internal): factor out common tests from em/remto pxto helper
